Renamed settings_display to settings_preferences
