+++ tools/diagnose
@@ -36,7 +36,7 @@ def check_python_version() -> bool:
 
 
 @run
-def pyPWD() -> bool:
+def pwd() -> bool:
     print(os.getcwd())
     return True
