+++ web/src/settings_realm_user_settings_defaults.js
@@ -5,7 +5,7 @@ import * as overlays from "./overlays";
 import {page_params} from "./page_params";
 import {realm_user_settings_defaults} from "./realm_user_settings_defaults";
 import * as settings_components from "./settings_components";
-import * as settings_display from "./settings_display";
+import * as settings_preferences from "./settings_preferences";
 import * as settings_notifications from "./settings_notifications";
 import * as settings_org from "./settings_org";
